Removed some of the more obvious errors. The remaining errors are mostly due to missing wrapper for operators or function overloading problems.

+++ copasi/bindings/python/unittests/Test_CTimeSeries.py
@@ -22,12 +22,12 @@ class Test_CTimeSeries(unittest.TestCase):
 
   def test_getData(self):
     data=self.ctimeseries.getData(0,1)
-    self.assert_(type(data)==DoubleType)
+    self.assert_(type(data)==FloatType)
     self.assert_(data==float("nan"))
 
   def test_getConcentrationData(self):
     data=self.ctimeseries.getConcentrationData(0,1)
-    self.assert_(type(data)==DoubleType)
+    self.assert_(type(data)==FloatType)
     self.assert_(data==float("nan"))
 
   def test_getTitle(self):
